feat(SKILL-170): drain telemetry outbox at CLI runtime completion boundaries
Drain the completion telemetry outbox via TelemetryService.autoSync at the
two runtime completion boundaries (feature-task runtime, goal). The drain
runs on a daemon thread bounded to 5s and swallows every exception, so exit
code, stdout, and reported outcome are unaffected. Telemetry level gating is
inherited from autoSync.

Adds reusable test support (RecordingTelemetryRequester,
CliCompletionTelemetryDrainSupport) and a call-site-removal mutation check
pinning both drain call sites.
