+++ runtime-kotlin/runtime-cli/src/test/kotlin/skillbill/cli/CliGoalRuntimeTest.kt
@@ -87,8 +87,8 @@ class CliGoalRuntimeTest {
     assertEquals(0, result.exitCode, result.stdout)
     assertContains(result.stdout, "goal SKILL-901: finished")
     assertEquals(0, pendingTelemetryOutboxCount(fixture.dbPath))
-    assertEquals(1, requester.requests.size, requester.requests.toString())
-    assertContains(requester.requests.single(), TELEMETRY_FIXTURE_PROXY_URL)
+    assertTrue(requester.requests.isNotEmpty())
+    assertTrue(requester.requests.all { it.contains(TELEMETRY_FIXTURE_PROXY_URL) }, requester.requests.toString())
   }
 
   @Test
@@ -1887,6 +1887,7 @@ internal data class GoalCliFixture(
     requester: HttpRequester = UnconfiguredHttpRequester,
   ): CliRuntimeContext = CliRuntimeContext(
     userHome = tempDir,
+    environment = emptyMap(),
     requester = requester,
     workflowGitOperations = workflowGitOperations,
     agentRunLauncher = launcher,
